refactor(console): market
