+++ console/src/layouts/Sidebar.tsx
@@ -41,6 +41,7 @@ import {
   SparkDebugLine,
   SparkSaveLine,
   SparkEmailLine,
+  SparkCardLine,
 } from "@agentscope-ai/icons";
 import { Package } from "lucide-react";
 import { clearAuthToken } from "../api/config";
@@ -473,7 +474,7 @@ export default function Sidebar({ selectedKey }: SidebarProps) {
         {
           key: "market",
           label: collapsed ? null : t("nav.market", "Skill Market"),
-          icon: <SparkOtherLine size={16} />,
+          icon: <SparkCardLine size={16} />,
         },
         {
           key: "environments",
