+++ console/src/pages/Settings/Market/components/DetailDrawer.module.less
@@ -1,0 +1,90 @@
+.detailHeader {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  margin-bottom: 16px;
+}
+
+.detailHeaderText {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.detailTitle {
+  margin: 0;
+  font-size: 18px;
+  font-weight: 600;
+  color: rgba(20, 20, 19, 0.88);
+}
+
+.detailMeta {
+  display: flex;
+  gap: 8px;
+  align-items: center;
+}
+
+.sourceBadge {
+  display: inline-flex;
+  align-items: center;
+  font-size: 11px;
+  padding: 1px 8px;
+  border-radius: 4px;
+  font-weight: 500;
+  background-color: rgba(239, 240, 243, 0.9);
+  color: #8e8c99;
+  border: 1px solid rgba(0, 0, 0, 0.06);
+  white-space: nowrap;
+}
+
+.detailDescription {
+  font-size: 13px;
+  color: rgba(20, 20, 19, 0.85);
+  line-height: 1.6;
+  margin-bottom: 16px;
+  white-space: pre-wrap;
+}
+
+.detailRows {
+  margin: 0 0 16px 0;
+  padding: 0;
+  display: grid;
+  grid-template-columns: 100px 1fr;
+  gap: 8px 12px;
+}
+
+.detailRow {
+  display: contents;
+}
+
+.detailKey {
+  font-size: 12px;
+  color: rgba(20, 20, 19, 0.55);
+  font-weight: normal;
+  margin: 0;
+  padding-top: 2px;
+}
+
+.detailValue {
+  font-size: 13px;
+  color: rgba(20, 20, 19, 0.88);
+  margin: 0;
+  word-break: break-all;
+}
+
+.drawerFooter {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+}
+
+.mono {
+  font-family: ui-monospace, SFMono-Regular, Menlo, monospace;
+  font-size: 11px;
+  background: rgba(239, 240, 243, 0.7);
+  padding: 1px 6px;
+  border-radius: 3px;
+  color: rgba(20, 20, 19, 0.88);
+}
