+++ console/src/pages/Settings/Market/components/SkillIcon.module.less
@@ -1,0 +1,17 @@
+.skillIcon {
+  width: 52px;
+  height: 52px;
+  border-radius: 10px;
+  background: #f7f5f3;
+  flex-shrink: 0;
+  object-fit: cover;
+  font-size: 28px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  overflow: hidden;
+}
+
+img.skillIcon {
+  display: block;
+}
