+++ console/src/pages/Settings/Market/components/TargetToggle.module.less
@@ -1,0 +1,4 @@
+.targetToggle {
+  display: inline-flex;
+  gap: 6px;
+}
